backend/django_back/requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.2
bcrypt==4.0.1
billiard==4.1.0
boto3==1.28.0
botocore==1.31.0
celery==5.3.1
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.0
decorator==4.4.2
Django==4.2.3
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-environ==0.10.0
django-grafana==0.2.0
django-prometheus==2.3.1
django-storages==1.13.2
django-timezone-field==5.1
djangorestframework==3.14.0
drf-yasg==1.21.6
gunicorn==20.1.0
idna==3.4
imageio==2.31.1
imageio-ffmpeg==0.4.8
inflection==0.5.1
jmespath==1.0.1
kombu==5.3.1
moviepy==1.0.3
mysqlclient==2.2.0
numpy==1.25.1
opencv-python==4.8.0.74
packaging==23.1
Pillow==10.0.0
proglog==0.1.10
prometheus-client==0.17.1
prompt-toolkit==3.0.39
pydub==0.25.1
PyJWT==1.7.1
PyMySQL==1.1.0
python-crontab==3.0.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
redis==4.6.0
requests==2.31.0
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
tqdm==4.65.0
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
vine==5.0.0
wcwidth==0.2.6
flower==2.0.0
django-cors-headers==4.2.0